Project 1 Completed
Project 1 utilizes inter-process communication through built-in Java multiprocess support. It is designed to simulate a CPU and memory working together to execute instructions from a passed-in file. There are 5 samples provided, and each can be run by using the instructions in the ReadMe. The summary is an explanation of implementation.